smarter dupe check and file trimming
If the user f3+i's a spawner and inports it into MCStacker, edits it, then exports back into a command, it was possible that some spawners would have non-matching data even though the contents was identical. This is likley due to MC stacker inverting the order that the Count byte and the tag object of an item. To fix this I manually remove and re-add the tag object of each item on an entity in the hands and armor.